Linux port: Fix UI resource loading and rendering
This commit fixes the UI system to properly load and render resources
on 64-bit Linux. The main issues were:

1. Binary file format compatibility (64-bit vs 32-bit):
   - Windows binary .idx/.rsc files use 32-bit integers for sizes/offsets
   - On 64-bit Linux, 'long' is 8 bytes vs 4 bytes on Windows
   - Changed all file I/O variables from 'long' to 'int32_t'
   - Fixed fread() calls to use sizeof(int32_t) instead of sizeof(long)
   - This fixed std::bad_alloc crash from reading wrong size values

2. Struct alignment for binary file parsing:
   - ImageHeader, SoundHeader, FlatHeader structs used 'long' members
   - This caused misalignment when parsing binary resource index files
   - Changed all struct members to explicit int32_t/int16_t types
   - This fixed garbled image names ('AIN_BG' instead of 'UI_MAIN_BG')

3. File path handling for Linux:
   - OpenResFile() used Windows backslashes and case-sensitive fopen()
   - Added FF_LINUX code path with:
     - Backslash to forward slash conversion
     - Case-insensitive file lookup via fopen_nocase()
     - Duplicate "art/" prefix stripping

Files modified:
- src/ui95/imagersc.h: ImageHeader struct uses int32_t/int16_t
- src/ui95/soundrsc.h: SoundHeader struct uses int32_t/int16_t
- src/ui95/flatrsc.h: FlatHeader struct uses int32_t
- src/ui95/cresmgr.cpp: Fixed file I/O, added Linux path handling
- src/ui95/chash.cpp: Added debug diagnostics for hash lookup
- CLAUDE.md: Documented all fixes and architecture notes

Result: Landing page now renders with actual content instead of black.

+++ src/ui95/flatrsc.h
@@ -1,6 +1,8 @@
 #ifndef _FLAT_RSC_H_
 #define _FLAT_RSC_H_
 
+#include <stdint.h>
+
 //
 // This class is TIED to the C_Resmgr class (which holds the actual sound data)
 //
@@ -8,6 +10,7 @@
 //
 
 // First item MUST be     short Type
+// NOTE: Must use int32_t for all "long" fields to match Windows 32-bit binary file format
 class FlatHeader
 {
 #ifdef USE_SH_POOLS
@@ -23,10 +26,10 @@ public:
     };
 #endif
 public:
-    long  Type;
-    char  ID[32];
-    long  offset;
-    long  size;
+    int32_t Type;         // Must be 32-bit for binary file compatibility
+    char    ID[32];
+    int32_t offset;       // Must be 32-bit for binary file compatibility
+    int32_t size;         // Must be 32-bit for binary file compatibility
 };
 
 class FLAT_RSC
